changed weight from int -> decimal

+++ SlaughterHouse/SlaughterHouse/Entities/Animal.cs
@@ -4,6 +4,6 @@ public class Animal
 {
    public int AnimalId {get; set;}
    public DateTime ArrivalDate {get; set;}
-   public int AnimalWeight {get; set;}
+   public decimal AnimalWeight {get; set;}
    // public string Species {get; set;}
 }
